add comments1.3

+++ oms_cms/backend/news/views.py
@@ -116,5 +116,6 @@ class DeleteComment(View):
 
     def get(self, request, pk, lang=None, category=None, post=None):
         comm = get_object_or_404(Comments, id=pk)
-        comm.delete()
+        if comm.user == request.user:
+            comm.delete()
         return redirect('new-detail', lang=lang, category=category, post=post)
